[11]  specify the action of the Intent to be Intent.ACTION_VIEW which is an Android system action and the value of it is android.intent.action.VIEW. Use Uri.parse() method to parse the string and build a Uri object. Use setData to pass in the Uri object and we use again the syntax sugar mentioned earlier which looks like we’re assigning the value directly to the field of Intent.
By using the implicit Intent, we can start the Activity not only inside our app but also the activity in other apps which makes inter-app communication and sharing possible. For instance, if you want to display a web page in your app, then you don’t need to implement a browser but instead using the system browser to open it.