[18] create Standard Launch mode of Activity:  Android uses back stack to manage the activity, when a new activity gets started, it will be pushed into the stack and at the top of the stack.
Start FirstActivity from FirstActivity which is odd and not normally used in real projects, but is a good example of the standard mode usage. Add the the code to print the current instance of the activity to easily tell from the logs that every time the button has been clicked, a new instance of FirstActivity will be created. Create N instances of FirstActivity and click Back btn N times to exit the app.