[19] create SingleTop Launch mode of Activity:  standard mode doesn’t make a lot of sense since if an activity is already at the top of stack, why create a new activity instance? If the top of the stack is already an instance of the activity, then the system will use the existing instance instead of creating a new one.
No matter how many times the button is clicked, there will be no more logs captured. This is because the current activity instance is already at the top of the stack and every time when starting FirstActivity, system will just reuse the activity at the top of the stack. Since there is only one instance of FirstActivity, press Back button once to exit the app.

However, if FirstActivity is not at the top of the stack, starting FirstActivity will still create a new instance. Let's take the following example into  account.