[20]  modify onCreate() method in FirstActivity to start SecondActivity on btn click, modify the onCreate() method in SecondActivity to start  FirstActivity on btn click
If FirstActivity is not at the top of the stack, starting FirstActivity will create a new instance. See the Logcat messages for the fact that system created two or more different instances of FirstActivity, because when SecondActivity tries to start FirstActivity, the top of back stack is SecondActivity, so a new instance of FirstActivity will get created. Press Back btn to go to SecondActivity, then press Back btn again for FirstActivity and again to exit the app.