[21]  create SingleTask Launch mode of Activity: this will create only one instance of Activity through the app. When starting this activity, the system will check the existence of this activity in the back stack, and, if there is an instance of it, system will pop all activities above out of stack and reuse this instance.
Rerun the app from FirstActivity, click the button to open SecondActivity, and then click the button there to open FirstActivity. From logs, it follows that when SecondActivity starts FirstActivity,  there is an instance of FirstActivity below SecondActivity, SecondActivity will pop and FirstActivity will be at the top of stack again, and thus FirstActivity’s onRestart() method and SecondActivity’s onDestroy() method will be called. Now, there is only one activity instance in the stack which is FirstActivity. Press Back Btn once to exit the app.