[22]  create SingleInstance Launch mode of SecondActivity: while FirstActivity and ThirdActivity are kept in the same back stack, a separate back stack will be created to manage SecondActivity (and this activity can be used by other apps); modify the onCreate() methods to print task id.
Rerun the app from FirstActivity, click the button to open SecondActivity, and then click the button there to open ThirdActivity. From Logs we see that task ID of SecondActivity is different from FirstActivity's and ThirdActivity's, which means that SecondActivity instance is in a separate back stack.

Press the Back button to have the FirstActivity shown up. Press Back Btn again to show SecondActivity, the third press will exit the app. Why? Since FirstActivity and ThirdActivity are in the same back stack, the Back Btn will first pop ThirdActivity's out of the stack, and the FirstActivity will be at the top now and show up. On pressing Back Btn, the FirstActivity be popped out, and the back stack gets empty (and removed), and another back stack with SecondActivity be used to show. Popping out the SecondActivity will cause back stack to empty, and the app will exit.