[25+] starting Activity  with Best Practice:   start the activity by building the intent and then use startActivity() or startActivityForResult(), use the intent to pass data between activities
If you’re not the author of SecondActivity and need to start SecondActivity, but you’re not sure about what params are needed to start SecondActivity,  then you need to either read SecondActivity or ask the author of this class, which is time consuming.

Solve the problem by using companion object and defined actionStart() method (all the methods defined in companion object can be used as Java static methods).In the actionStart() method, create an intent instance and save the required data into the intent,then use startActivity() method to start SecondActivity. The benefit? The clarity! All params needed are there in the method params (without reading the code of SecondActivity nor asking the author). Also, this approach simplifies the code to start activity as we only need one line of code to start SecondActivity :

SecondActivity.actionStart(this, "data1", "data2")

This approach will help reduce the time to read and write code and questions from your colleague.