[45] create a toolbar having Back and Edit buttons and a TextView,  use Include Layout to include the toolbar in the current activity layout ( ui_layout.xml ) with just one line of code, as simple as that!
More attributes to be used: Android:background is used to set the background of the widget which can be a color or an image.  Android:layout_margin attribute can specify the margins at the four directions. Also use android:layout_marginLeft or android:layout_marginTop to set the margin at specific direction.

Make sure to hide the native system toolbar from activity: use getSupportActionBar() to get the instance of the ActionBar and then call hide() on it to hide itself. Since ActionBar is nullable, we need to use? . operator.