[46] create a Customized Widget: create TitleLayout that inherits LinearLayout, add this customized widget to ui_layout.xml, use the fully-qualified name:  TitleLayout, plus it's package name.
Reason: if the widget in the layout needs to respond to certain events, we still need to register the event handler in each of the activities. For instance, the back button in the toolbar only does one thing, which is to destroy the current activity. It is not a good idea to duplicate the event handling code in every activity. We can use the customized widget to solve this problem.

We declare context and attrs as the params of the primary constructor and when TitleLayout gets created, this primary constructor will be called.

The from() method of LayoutInflater in the init will create an instance of LayoutInflater and its inflate() method will dynamically mount a layout.  Inflate() takes two params: the first param is ID of the layout file, and we pass R.layout.title to it; the second param is the parent layout of the mounted layout, and we pass this to set it to the TitleLayout.