[50] optimize the efficiency of the ListView - step 1: the current implementation of ListView is slow because in getView() of FruitAdapter, the item view’s layout gets inflated separately for every item view and this will drag the performance of the app when the ListView is fast scrolling.
There is a convertView param in getView(). This param can be used to cache the inflated view and reuse it later which can be used to optimize the performance of ListView.

In getView(), if convertView is null, then use LayoutInflater to inflate the layout; if not, then reuse the convertView. By reducing the inflation action to once, fast scrolling will be faster and smoother.