[51] optimize the efficiency of the ListView - step 2:  every time getView() gets called, findViewById() method will get called to get reference of the view.  Use the inner class ViewHolder to optimize this, to cache the ImageView and TextView instance.
The setTag() method of View will set the tag to be the instance of ViewHolder. When convertView is not null, getTag() will get the instance of ViewHolder. This instance has the instances of TextView and ImageView. Thus there is no need to use findViewById() again.

After these two steps of optimization, this customized ListView will perform much more efficiently.