[52] create a click event in ListView
The setOnItemClickListener() method will register a listener for ListView which will execute the Lambda expression, when a user clicks any item in the ListView. The position param indicates, which item user is clicked and can be used to get the corresponding instance of the item. A toast message will show the fruit name.