[53] replace unused params in Lambda expression with underscores
The Lambda expression  declared four params. How do we know what params we need to declare? Use a shortcut CTRL+click to see the source code, and find that setOnItemClickListener() takes a param of OnItemClickListener which is a Java functional interface.

Although the four params must be declared in Lambda expression, only one (position) has been used. In Kotlin, underscore can be used to replace the unused params (the order of the params must be maintained).