[54] create a RecyclerView, a powerful widget providing more functionality than the ListView. Add the dependency to app/build.gradle file, update activity_main.xml, create FruitAdapter class that inherits RecyclerView.Adapter and set the param type to be FruitAdapter.ViewHolder (the inner class).
This is the standard way to create adapter for RecyclerView: ViewHolder is an inner class that inherits from RecyclerView.ViewHolder. It takes a View param in the primary constructor which usually is outer layout of the item view in RecyclerView. The instances of ImageView and TextView can be acquired by findViewById(). The data source is passed in the primary constructor of FruitAdapter. FruitAdapter inherits RecyclerView.Adapter. Methods onCreateViewHolder(), onBindViewHolder() and getItemCount() need to be overridden.

onCreateViewHolder() is used to create instance of ViewHolder. fruit_item layout will be inflated then ViewHolder will be instantiated with the inflated view.

onBindViewHolder() will set value for the item which will be executed when the item has been scrolled into the screen. Use position to get the instance of the current item, then use the item to set image, and text for ImageView, and TextView in ViewHolder.

An instance of LinearLayoutManager gets created in onCreate(), and gets used to specify the RecyclerView’s layoutManager getItemCount() returns the size of the data set.

LinearLayoutManager will position the items like ListView. Next, we instantiate the FruitAdapter with the fruits data and assign this adapter to the RecyclerView instance to bridge RecyclerView and data.

The logic of RecyclerView should be clearer compared with ListView.