[56] experiment with a fancier StaggeredGridLayoutManager to show waterfall flow layout.
In fruit_item.xml, change the width of the LinearLayout from 80 dp to match_parent since the width of the item view should be determined by the column number of the layout instead of a fixed value.

Use layout_margin to give some space between the item views.

Change the layout_gravity of TextView to left because we make the text longer making left alignment more sensible.

In ListViewActivity.kt, the StaggeredGridLayoutManager constructor takes two params: the first param is the columns of the layout and there are three columns in this example; the second param will specify the orientation of the layout which will be vertical.

Besides LinearLayoutManager, LayoutManager has two other implementations supported natively: GridLayoutManager and StaggeredGridLayoutManager.