[60] build a beautiful Chat User Interface
There are only two fields in Msg class to store message content and message type. Message type has two values: TYPE_RECEIVED and TYPE_SENT. (const can only be used in singleton, companion object, or top-level method.)

Initialize data in initMsg().  The send button event listener gets the content of EditText view and if the content is not empty string, a new instance of Msg will be created and added to msgList.  The notifyItemInserted() method will notify the change and refresh the RecyclerView to show the latest message.

(You can also use the notifyDataSetChanged() method of
 the adapter which will refresh all the visible elements in the RecyclerView. Thus  whether it is insertion, deletion, or updating, UI will show the latest data, but this is relatively slow)

The scrollToPosition() method will scroll
the RecyclerView to the last item. This is to make sure the latest message will appear.

The setText() method clears the EditText view.