[61] use lateinit keyword for late initialization. This eliminates the need to add nullability check in onClick(), when the number of global variables is growing.
The adapter is declared as global variable but gets initialized in onCreate(). Thus adapter was initialized with value of null and its type is MsgAdapter?.  We initialize the adapter in onCreate(). However compiler will fail to compile if we don’t add null check in onClick().

As you add more and more global variables, this problem will become more and more serious. You probably will need to write lots of null check code just to make the code can compile.

One of the solutions is using lateinit. This keyword tells Kotlin compiler that the variable declared with it will be initialized later and there is no need to assign null to the variable when gets declared.

With the lateinit keyword, there is no need to assign null to the variable and the variable type becomes MsgAdapter instead of MsgAdapter?. This eliminates the need to add nullability check in onClick().

However, if adapter gets referenced before initialized, then program will crash and throw UninitializedPropertyAccessException.

Make sure that the lateinit variables are initialized before using them to ensure null safety of the program.

We can check if a global variable has been initialized or not to avoid the exception above and unnecessary initialization. ( ::adapter.isInitialized determines if adapter has been initialized)