[62]  optimize with Sealed Class: create sealed class  MsgViewHolder and two subclasses
There is no need to add else statement in the when(holder) {...} condition check of the onBindViewHolder method.

When the when statement gets a sealed class as argument, Kotlin compiler will check what are the sub classes that inherits this class, then check if all the cases are covered which can eliminate the else statement.

If we add Unknown class that inherits MsgViewHolder without adding the corresponding condition check in onBindViewHolder(), compiler will fail to compile.

The sealed class and its sub classes can only be defined at the top level of the same file and cannot be embedded in other classes which is a limitation of the sealed class’s implementation.