[65] add Back Stack for Fragment : we want the previous Fragment to show up instead of exiting the  app just like the Activity back stack.
FragmentTransaction has addToBackStack() to add a transaction to back stack.

Before the transaction gets committed, it gets added to the back stack by calling addToBackStack() which takes a string as param. This string is the name for the back stack state and is optional. Usually, we just pass null for the name.

Click the Back Btn. App will not exit but go back to the RightFragment screen. Click Back button again, RightFragment will disappear too. After another click, app will exit.