[69] create a basic News App :  create News class for news, create layout file news_content_frag.xml to host the content of the news,
There are two main parts of the content layout: title at the top and article at the bottom. Between them, we use a horizontal thin line as divider. We also use a vertical think line to divide the news list and the news content. To make a thin line, we just need to create a View and set its width or height to have value of 1 dp and use background attribute to set the color of the line. Here we set the color of the thin line to be black.

Set the content layout to be invisible since, in two-pane mode, this layout should not appear without selecting news.

Create NewsContentFragment class that inherits Fragment.  Inflate the news_content_frag layout in onCreateView(). refresh() will show the title and content of news.

Create NewsContentActivity and activity_news_content.xml.

Create news_title_fragment.xml (the layout that can display the list of news titles)

Create news_item.xml as the item view layout for RecyclerView.

After creating the list and item view layouts, we need a place to display them: create NewsTitleFragment to be the Fragment that will host the news title list.

Create layout-sw600dp folder and under this folder create activity_fragment.xml. In two-pane mode, we have two Fragments and place the Fragment for news  content in a FrameLayout whose id is newsContentLayout. Thus, if this view can be referenced then display mode is two-pane otherwise one-pane.

One last thing to display the news list in the RecyclerView is to create the adapter. We can create an inner class NewsAdapter for the RecyclerView

The click event listener in onCreateViewholder() first gets the News instance at the clicked position, then use isTwoPane to decide the display mode. If it is one-pane mode, then start a new Activity to display the content; if it is two-pane mode, then refresh NewsContentFragment with the new data.

We're getting the NullPointerException because we're trying to call fragment?.refresh(title, content) before the NewsContentFragment's view hierarchy has been fully created and its binding initialized. Let's break down the problem and explore the best solutions.  Here's a more detailed explanation of what's happening:

1. NewsContentActivity's onCreate(): Our NewsContentActivity's onCreate() method is called when the activity is created.

2. Fragment Creation: Inside NewsContentActivity's onCreate(), we're doing the following:

* Retrieving data from the Intent (title and content).
* Finding the fragment using supportFragmentManager.findFragmentById(R.id.newsContentFrag).
* Attempting to call fragment?.refresh(title, content).

3.Fragment Lifecycle: The problem is that the NewsContentFragment has its own lifecycle, which is asynchronous to the activity's lifecycle. Here's what happens:
* onCreate(): The fragment's onCreate() is called.
* onCreateView(): The fragment's onCreateView() is called. This is where the fragment's layout is inflated, and the binding is initialized.
* onViewCreated(): The fragment's onViewCreated() is called. This is called after onCreateView() and is the first moment when the fragment's view hierarchy is fully created and ready to be used.

4. The Race Condition: The issue is that our NewsContentActivity's onCreate() is likely finishing before the NewsContentFragment's onCreateView() and onViewCreated() have been called. This means that when we call fragment?.refresh(title, content), the fragment's view hierarchy (and therefore its binding) is not yet ready, and we get a NullPointerException when refresh() tries to access the binding.

Solution :

 Move refresh() Call to onViewCreated() (Recommended) This is the most robust and recommended solution. It involves moving the logic that depends on the fragment's view hierarchy to the fragment's onViewCreated() method.

5. Pass Data via arguments: Instead of calling refresh() directly, pass the title and content to the fragment using arguments when you create the fragment.

6. Retrieve Data in onViewCreated(): In the fragment's onViewCreated(), retrieve the data from arguments and then perform the refresh logic.

 Explanation :

 * NewsContentActivity
 - The refresh() call is removed.
 - A Bundle is created to hold the title and content.
 - The Bundle is set as the fragment's arguments.

 * NewsContentFragment:
 - The refresh() method remains, but it's no longer called from the activity.
 - onViewCreated() is used to retrieve the data from arguments.
 - refresh() is called inside onViewCreated(), ensuring the view is ready.

 * Correct Timing: onViewCreated() is guaranteed to be called after the view hierarchy is created.
 * Clean Separation: The fragment is responsible for its own data and view updates.
 * No Race Conditions: This approach eliminates the race condition.
 * arguments : This is the correct way to pass data to a fragment.
 * onViewCreated() : This is the correct place to access the view.

 That is everything we need to do.

 Let us run it in the emulator! First run the app in  phone emulator. We see lots of news titles and clicking the title of the news will start a new Activity to show the news details.

 Next, run the app in tablet emulator and click the news title(to change the title and the content on the right hand-side).

 As we can see, with the same code, the UI will appear differently in the phone and the tablet.