[72] define a  BroadcastReceiver to receive a broadcast: when MyBroadcastReceiver receives the customized broadcast, it will show a toast with message “received in MyBroadcastReceiver.” , edit AndroidManifest.xml to make MyBroadcastReceiver to receive the broadcas with value of com.example.broadcasttest.MY_BROADCAST, add a button in activity_broadcast_test_main.xml, set OnClickListener for this btn to trigger sending the broadcast:
First, we need to create an Intent instance and pass the value of the broadcast we want to send.

Use the setPackage() method of Intent to pass in the package name of the current app. As mentioned before, after Android 8.0, static BroadcastReceivers cannot receive implicit broadcasts and by default all of our customized broadcasts are implicit broadcasts. Thus we need to call setPackage() to specify which app is going to receive this broadcast to make it an explicit broadcast, otherwise the static BroadcastReceiver won’t be able to receive this broadcast.

Finally, use sendOrderedBroadcast(intent, null ), where the 1st param is the Intent type and the 2nd param is the permissions string, [or sendBroadcast(intent)] to send ordered [normal]  broadcast, and all the BroadcastReceivers listening to com.example.broadcasttest.MY_BROADCAST