[77]  persist/load Data in/from File: Context class provides openFileOutput() to store the data in the specified file. This method will take two params. The first param is the file name.  The second param is the operation mode to the file which has two values: MODE_PRIVATE and MODE_APPEND. The default is MODE_PRIVATE, which means that the new data will overwrite the original file’s content if it exists. MODE_APPEND means that the new data will be appended at the end of the file if it exists, if not, a new file will be created.
In the save method, we get an object FileOutputStream through openFileOutput() , and use it to construct OutputStreamWriter, and use the created object to construct the BufferedWriter. The BufferedWriter uses internal extension function 'use', which will automatically close the stream of outer layer after the outer layer of the Lambda expression. Thus there is no need for us to write the finally statement to close the stream manually.

Create activity_file_persistence_test.xml with  EditText view for text input.

Override onDestroy() to make sure that save() can be called before Activity gets destroyed.

Open  Device File Explorer and browse to /data/data/com.example.activitytest/files/data  and open the data file to see that the content in the EditText view has been successfully stored in the data file.

Now, read the Data from data file. In the load method, the ContextWrapper class provides the  openFileInput() to read data from file. This method is simpler compared with openFileOutput() as it only takes one param, which is the file name that will be read, and then system will go to /data/data/ <package name>/files/ folder to load this file and return an instance of FileInputStream.  Then use this instance to create an instance of InputStreamReader; after that, use this instance to create an instance of BufferedReader, with this object we can read the data in the file line by line and append them into the StringBuilder instance.

Notice that Kotlin internal extension function forEachLine is used here to pass each line in the file to the Lambda expression, which has the appending logic.