[80] create Database : Android provides SQLiteOpenHelper class to help us manage the database. There are two very important methods in SQLiteOpenHelper: getReadableDatabase() and getWritableDatabase(). Both of these two methods can  create or open an existing database (if database exists then open, if not then create a new one), then return an object that can read and write.
create MyDatabaseHelper extending SQLiteOpenHelper. There are two constructors in SQLiteOpenHelper. Usually,  we can override  the constructor and use the one with less param.   This constructor being  overridden takes four params: the first param is Context which is essential to operate on the database; the second param specifies the name of the database; the third params allows us to return a customized Cursor when we query the data and usually we just pass null; and the fourth param is to specify the version number of the current database.

The data-base files are created (on Btn click) under /data/data/<package name>/databases/.