[87] create a transaction :  SQLite database supports transaction which can ensure that a series of operations to be unit operation, which means that these operations either all finish or none.
We want to replace Book table with new data, we use delete() to delete the old data from Book table, and then use insert() to add the new data into the table. Make sure that deletion and adding data can succeed or fail altogether.

Here's  the standard way to do transaction in Android:

We use beginTransaction() in SQLiteDatabase to start a transaction, then we execute the database business logic in try and catch block. After everything finishes, we call setTransactionSuccessful() to indicate that transaction succeeds and finally call endTransaction() in finally block to end the transaction.

After deletion operation, we manually throw a NullPointerException so that adding data will not be called. However, since these operations are in a transaction, the old data should not be deleted.

Run the app and click the Replace Data btn to see that  Book table still only has the old data as expected.