[90] simplify Use of SharedPreferences (1) and Simplify Use of Content Values (2): use higher-order functions to simplify SharedPreferences and ContentValues.
(1) :  we directly call "edit" on the object of SharedPreferences and add data in the Lambda expression. Notice that Lambda expression now has the context of SharedPreferences.Editor, thus we can directly call put to add data. We no longer need to call apply to commit the action because "edit" will do it (KTX extension).

(2) :   Note the use of mapOf() function which  creates  key value pairs like "Apple" to 1, "Orange" to 2, etc... Create the  cvOf() method taking  params of Pair type  (vararg keyword is corresponding to variable arguments in Java which allows us to pass 0,1,2 or any number of arguments of Pair type). Pair is a data structure of key value pair, thus we need to use generics to specify the type of key and value.  All  keys are of string type in ContentValues, and we can specify the key of Pair to String. But since the value of ContentValues has multiple types to choose from (integer, float, string, or even null), we need to set the value to be type of Any?. This is because Any is the base class of all classes in Kotlin which is corresponding to Object in Java. Any? means that null value is allowed.

The basic idea of cvOf() function is that we create an object of ContentValues and iterate the pair argument list to get the value and add into ContentValues object and return it. Use when statement to check all the supported data types of ContentValues. Use for-in loop to iterate the  pairs argument lists and get the key and value in the loop and uses when statement to check the type of value. All the supported data types of ContentValues are covered.  The values are added into ContentValues one by one and in the end return the object of ContentValues. Here, Kotlin also applies Smart Cast. For instance, here for the code that is in the Int condition block, value will be automatically cast to Int type instead of Any?. Thus, we do not need to do type casting like Java. This applies to if statement.

With cvOf(), it is easier to use ContentValues. For example, to add data to database, we can have neat code as mentioned, and  we can use syntax like mapOf() to construct the ContentValues object, which is great!

The cvOf() is not the  higher-order function, but can be simplified to the higher-order function, for instance with apply function.