[94] read System Contact: create two new contacts in the Emulator manually. Create layout activity_contacts_test.xml and use ListView instead of RecyclerView for simplicity. In onCreate(), initialize ListView using the standard approach, then  start to handle the runtime permission request logic.  This is because READ_CONTACTS is a runtime permission. If we can get the permission, then we will call readContacts() to read the contacts’ information in system Contact app.
In readContact(), use query() in ContentResolver to query the contact data.  ContactsContract.CommonDataKinds.Phone encapsulates and provides a CONTENT_URI constant, which is the result of Uri.parse().  Then, iterate over the Cursor object returned by the query() method, using the ?. operator and the apply function to simplify the traversal code.

In apply function,  get the contact name and phone number one by one. The column for contact name is ContactsContract.CommonDataKinds.Phone.DISPLAY, and the column for contact phone number is ContactsContract.CommonDataKinds.Phone.NUMBER. Concatenate the results and between them we add a line break, then add the concatenated data to ListView’s data source and refresh the ListView.

At the end, do remember to close the Cursor object. Modify AndroidManifest.xml  to declare the permission for android.permission.READ_CONTACTS for reading the system Contact.

Run the app to have a dialog popping up to request permissions to access the contacts, click Allow to have the list populated with contacts.