[95] implement Notification: add the notification permission and build the logic to send a basic notification. Create the NotificationTestActivity.
Add the `POST_NOTIFICATIONS` permission in `AndroidManifest.xml`. Create a layout named `activity_notification_test.xml` containing a button that can trigger a notification. Create the `NotificationTestActivity` to send a basic notification. The activity checks for Android version. If it is Android 8.0 or higher, then it will create a notification channel. Add a click listener on the button, which will build a simple notification and send it. Include small and large icons for the notification, and set the title and text. Add a new `large_icon.png`, `small_icon.png` and `big_image.jpg` in `res/drawable-xxhdpi`. Change the app name to "NotificationTest". Change the main activity to `NotificationTestActivity`.